Refactor sound factory and add Goriya hit sound
