+++ cse3902/Entities/Enemies/Goriya.cs
@@ -6,6 +6,7 @@ using cse3902.SpriteFactory;
 using cse3902.Sprites.EnemySprites;
 using Microsoft.Xna.Framework;
 using cse3902.Constants;
+using cse3902.Sounds;
 
 namespace cse3902.Entities.Enemies
 {
@@ -69,7 +70,7 @@ namespace cse3902.Entities.Enemies
 
         public void TakeDamage(int damage)
         {
-
+            SoundFactory.PlaySound(SoundFactory.Instance.enemyHit);
             this.Health -= damage;
             //this.goriyaSprite.Damaged = true;
             this.collidable.DamageDisabled = true;
